requirements.txt
spotipy
psycopg2-binary
python-dotenv
firebase-admin
